requirements/requirements.txt
aiohttp==3.8.3
black==22.12.0
flake8==6.0.0
fqdn==1.5.1
isoduration==20.11.0
isort==5.11.4
jsonpointer==2.3
jupyter==1.0.0
mypy==0.991
pandas==1.5.3
path==16.6.0
pip-chill==1.0.1
pre-commit==2.21.0
pydantic==1.10.4
pytorch-lightning==1.9.0
scikit-learn==1.2.1
torchaudio==0.13.1
torchvision==0.14.1
uri-template==1.2.0
webcolors==1.12
